refactor: 未使用の date/time フォーマッタを削除し format.ts を duration.ts へ改名
production 未使用（テストのみ参照）だった formatDate / formatDateShort /
formatDateISO / formatTime / formatHour / formatTimeRange / formatDuration(ms) を
削除し、唯一生きている formatDurationMinutes だけを残して format.ts を
duration.ts へ改名する。これで formatDate / formatTime / formatTimeRange の
同名衝突が解消する（日付・時刻の実表示は useDateFormat hook が担う）。

- calendar dateHelpers の未使用 formatDate を削除
- timeblock/lib/formatters.ts（全4関数 未使用）をテストごと削除
- 上記削除で orphan 化した MS_PER_HOUR と barrel の未使用再export を除去
- barrel は formatDurationMinutes を公開、TimeblockRelationshipSection を barrel import へ
- timezone.ts の未実装 formatInTimezone を指す stale JSDoc を修正

Refs #1525
