+++ apps/product/src/lib/date/__tests__/duration.test.ts
@@ -1,0 +1,25 @@
+import { describe, expect, it } from 'vitest';
+
+import { formatDurationMinutes } from '../duration';
+
+describe('date/duration', () => {
+  describe('formatDurationMinutes', () => {
+    // 共通入力テーブル（整数）— duration formatter 統合後も canonical はこの契約を維持する。
+    // 統合対象（formatDurationDisplay / formatMinutesDuration / formatMetricValue / 各 local）は
+    // 整数入力でこの出力に一致する。
+    it.each([
+      [0, '0m'],
+      [1, '1m'],
+      [45, '45m'],
+      [59, '59m'],
+      [60, '1h'],
+      [61, '1h 1m'],
+      [90, '1h 30m'],
+      [119, '1h 59m'],
+      [120, '2h'],
+      [1439, '23h 59m'],
+    ])('%i分 → %s', (input, expected) => {
+      expect(formatDurationMinutes(input)).toBe(expected);
+    });
+  });
+});
