+++ apps/product/src/lib/date/index.ts
@@ -29,9 +29,9 @@ export {
 } from './core';
 
 // ========================================
-// Format - フォーマット
+// Duration - 期間フォーマット
 // ========================================
-export { formatTime, formatTimeRange } from './format';
+export { formatDurationMinutes } from './duration';
 
 // ========================================
 // TimeString - "HH:mm" パース / フォーマット
@@ -41,4 +41,4 @@ export { formatHHmm, formatTimeString, parseTimeString } from './timeString';
 // ========================================
 // 定数の再エクスポート
 // ========================================
-export { CACHE_5_MINUTES, MS_PER_DAY, MS_PER_HOUR, MS_PER_MINUTE } from './constants';
+export { CACHE_5_MINUTES, MS_PER_MINUTE } from './constants';
